+++ utils/user.py
@@ -1,0 +1,2 @@
+def get_user_id(user: dict) -> str:
+    return user.get("sub")
